requirements.txt
django-cleanup
boto3
